requirements.txt
# Environment
python-dotenv>=1.0.0

# Azure SDKs
azure-storage-blob>=12.19.0
azure-core>=1.29.0

# OpenAI
openai>=1.14.0

# PDF extraction
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Text processing
markdown>=3.5.0
nltk>=3.8.0
langchain-text-splitters>=0.3.6
# Optional: local embeddings (dev / CI)
sentence-transformers>=2.6.0
faiss-cpu>=1.8.0
rank-bm25>=0.2.2

# Jupyter
jupyter>=1.0.0
ipykernel>=6.29.0
